+++ src/components/CoreConcept.jsx
@@ -1,0 +1,20 @@
+// function CoreConcept(props) {
+//   return (
+//     <li>
+//       <img src={props.image} alt={props.title} />
+//       <h3>{props.title}</h3>
+//       <p>{props.description}</p>
+//     </li>
+//   );
+// }
+
+// Using Object De-structuring
+export default function CoreConcept({image, title, description}) {
+  return (
+    <li>
+      <img src={image} alt={title} />
+      <h3>{title}</h3>
+      <p>{description}</p>
+    </li>
+  );
+}
